Add own system/extras repo to suppress EMMC-corrupting wipe command
If BOARD_SUPPRESS_EMMC_WIPE is true, the EMMC wipe command will not be
issued. This works around a bug in some firmware revisions of Samsung EMMC's
that permanently damages the device when the wipe command is issued.

This is only part of the solution but it does handle the "flashing CM9
for the first time on an unsafe kernel" situation.